Working JSON Generator parse and style format
This upload contains the working system for generating file output from a developed scraping interface style to a standardized and concatenated JSON style file. This also contains the scraping system used for testing through the use of the rocking r bar scraping system. This update lacks accurate file paths for each python file due to not knowing the overall path for the server